requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
motor==3.7.1
pymongo==4.13.2
bcrypt==4.3.0
python-jose[cryptography]==3.5.0
python-dotenv==1.1.1
razorpay==1.4.2
pydantic[email]==2.11.7
email-validator==2.2.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2 
setuptools 